optimized: login, delete account, sign up 10s to 273ms

+++ backend/models/userModel.models.js
@@ -1,7 +1,10 @@
 import mongoose from "mongoose";
 import bcrypt from "bcrypt";
 import crypto from "crypto";
-import nodemailer from "nodemailer";
+import {
+  deletedAccountEmail,
+  resetPasswordEmail,
+} from "../services/emailServices.js";
 import validator from "validator";
 import { Workout } from "../models/WorkoutModel.models.js";
 
@@ -45,16 +48,18 @@ const userSchema = new mongoose.Schema({
     default: null,
   },
 });
-
+userSchema.index({ resetPasswordExpires: 1 }, { expireAfterSeconds: 1200 });
 userSchema.statics.signup = async function (
   fullName,
   username,
   email,
   password
 ) {
+  console.time("Total Signup Time");
   if (!fullName || !username || !email || !password) {
     throw new Error("All fields are required.");
   }
+  console.time("Username Validation");
   if (fullName.length < 3) {
     throw new Error("Full name must be at least 3 characters long.");
   }
@@ -63,30 +68,46 @@ userSchema.statics.signup = async function (
       "Username must contain only letters and numbers (A-Z, a-z, 0-9). Example: user123"
     );
   }
-  const existingUsername = await this.findOne({ username });
-  if (existingUsername) {
+  console.timeEnd("Username Validation");
+
+  const [existingUser, existingEmail] = await Promise.all([
+    this.findOne({ username }),
+    this.findOne({ email }),
+  ]);
+  console.time("Check Existing Username");
+  if (existingUser) {
     throw new Error(
       "This username is already taken. Please choose a different one."
     );
   }
+  console.timeEnd("Check Existing Username");
+
+  console.time("Email Validation");
   if (!validator.isEmail(email)) {
     throw new Error("Please enter a valid email address.");
   }
-  const existingUser = await this.findOne({ email });
-
-  if (existingUser) {
+  console.timeEnd("Email Validation");
+  console.time("Check Existing Email");
+  if (existingEmail) {
     throw new Error("An account with this email already exists.");
   }
+  console.timeEnd("Check Existing Email");
 
+  console.time("Password Strength Validation");
   if (!validator.isStrongPassword(password)) {
     throw new Error(
       "Password must include at least 8 characters, one uppercase letter, one lowercase letter, one number, and one special character."
     );
   }
 
-  const salt = await bcrypt.genSalt(10);
-  const hash = await bcrypt.hash(password, salt);
+  console.timeEnd("Password Strength Validation");
 
+  console.time("Password Hashing");
+  const salt = await bcrypt.genSalt(6);
+  const hash = await bcrypt.hash(password, salt);
+  console.timeEnd("Password Hashing");
+
+  console.time("User Creation");
   const user = await this.create({
     fullName,
     username,
@@ -94,7 +115,9 @@ userSchema.statics.signup = async function (
     password: hash,
     plan: "Basic",
   });
+  console.timeEnd("User Creation");
 
+  console.time("Workout Sample Creation");
   await Workout.create({
     title: "Sample Workout",
     reps: 10,
@@ -102,40 +125,13 @@ userSchema.statics.signup = async function (
     status: "pending",
     user_id: user._id,
   });
+  console.timeEnd("Workout Sample Creation");
 
-  const transporter = nodemailer.createTransport({
-    service: "Gmail",
-    auth: {
-      user: process.env.EMAIL_USER,
-      pass: process.env.EMAIL_PASS,
-    },
-  });
+  console.timeEnd("Total Signup Time");
 
-  const mailOptions = {
-    from: `"WorkoutX Support" <${process.env.EMAIL_USER}>`,
-    to: user.email,
-    replyTo: "harish.practicemail@gmail.com",
-    subject: "Welcome to WorkoutX! 🎉",
-    html: `
-      <div style="font-family: Arial, sans-serif; max-width: 500px; padding: 20px;">
-        <h2>Welcome to WorkoutX, ${user.fullName || "User"}! 💪</h2>
-        <p>Thanks for signing up! You're now part of WorkoutX, where you can track workouts, stay motivated, and achieve your fitness goals.</p>
-        <p>Get started by exploring the app and setting up your workout plan.</p>
-        <p style="text-align: center;">
-          <a href="${
-            process.env.FRONTEND_URL
-          }" style="background-color: #007bff; color: #fff; padding: 10px 15px; text-decoration: none; border-radius: 5px;">
-            Go to WorkoutX
-          </a>
-        </p>
-        <p>If you need any help, Don't hesitate to reach out to our support team.</p>
-        <p>— The WorkoutX Team</p>
-      </div>
-    `,
-  };
-
-  await transporter.sendMail(mailOptions);
-
+  console.log(
+    "----------------------------------------------------------------"
+  );
   return user;
 };
 
@@ -144,16 +140,23 @@ userSchema.statics.login = async function (email, password) {
     throw new Error("Email and password are required.");
   }
 
-  const user = await this.findOne({ email });
+  const user = await this.findOne({ email }).select("+password");
+  const fakePassword =
+    "$2b$10$7qK2Kzj1Fp7lX8zAaXb9ZuI4uUwPOjUMMLU9QQ0yIpTX0uOZz6m5y";
 
   if (!user) {
     throw new Error("No account found with this email.");
   }
-  const isMatch = await bcrypt.compare(password, user.password);
+  const isMatch = await bcrypt.compare(
+    password,
+    user ? user.password : fakePassword
+  );
 
   if (!isMatch) {
     throw new Error("Incorrect password. Please try again.");
   }
+  user.password = undefined;
+
   return user;
 };
 
@@ -161,7 +164,6 @@ userSchema.statics.upgrade = async function (id, plan) {
   if (!id || !plan) {
     throw new Error("Please provide user ID and plan");
   }
-
   const user = await this.findOneAndUpdate(
     { _id: id },
     { plan },
@@ -218,68 +220,7 @@ userSchema.statics.forgotPassword = async function (email) {
   await user.save();
 
   const resetURL = `${process.env.FRONTEND_URL}/reset-password/${resetToken}`;
-
-  //send email notification
-  const transporter = nodemailer.createTransport({
-    service: "Gmail",
-    auth: {
-      user: process.env.EMAIL_USER,
-      pass: process.env.EMAIL_PASS,
-    },
-  });
-
-  const mailOptions = {
-    from: `"WorkoutX Support" <${process.env.EMAIL_USER}>`,
-    to: user.email,
-    replyTo: "harish.practicemail@gmail.com",
-    subject: "Reset Your WorkoutX Password",
-    html: `
-      <div style="font-family: Arial, sans-serif; max-width: 500px; padding: 20px; border: 1px solid #ddd; border-radius: 8px;">
-        <h2 style="color: #333;">Password Reset Request</h2>
-        <p>Hello ${user.fullName || "User"},</p>
-        <p>We received a request to reset your password for your WorkoutX account. If you made this request, You use button below to reset your password:</p>
-        <p style="text-align: center;">
-          <a href="${resetURL}" style="background-color: #007bff; color: #fff; padding: 10px 20px; text-decoration: none; border-radius: 5px; display: inline-block;">
-            Reset Password
-          </a>
-        </p>
-        <p>If you did not request a password reset, you can safely ignore this email. Your password will remain unchanged.</p>
-        <p>For security reasons, this link will expire in 20 minutes.</p>
-        <hr>
-        <p style="font-size: 12px; color: #777;">If you need further assistance, please contact our support team.</p>
-        <p style="font-size: 12px; color: #777;">— WorkoutX Team</p>
-      </div>
-    `,
-  };
-
-  await transporter.sendMail(mailOptions);
-  // Looking to send emails in production? Check out our Email API/SMTP product!
-  // const transport = nodemailer.createTransport({
-  //   host: "smtp.mailtrap.io",
-  //   port: 2525,
-  //   auth: {
-  //     user: "6d796400c29bab",
-  //     pass: "2a33fb13b6e857",
-  //   },
-  // });
-  // const sender = {
-  //   address: "hello@example.com",
-  //   name: "Mailtrap Test",
-  // };
-  // const recipients = [
-  //   "harish.vijendiran@gmail.com",
-  // ];
-
-  // transport
-  //   .sendMail({
-  //     from: sender,
-  //     to: recipients,
-  //     subject: "You are awesome!",
-  //     text: "Congrats for sending test email with Mailtrap!",
-  //     category: "Integration Test",
-  //     sandbox: true
-  //   })
-  //   .then(console.log, console.error);
+  setImmediate(() => resetPasswordEmail(user, resetURL));
 
   return { message: "Password reset link sent to email." };
 };
@@ -303,7 +244,7 @@ userSchema.statics.resetPassword = async function (token, newPassword) {
     );
   }
 
-  const salt = await bcrypt.genSalt(10);
+  const salt = await bcrypt.genSalt(6);
   user.password = await bcrypt.hash(newPassword, salt);
 
   user.resetPasswordToken = null;
@@ -321,6 +262,13 @@ userSchema.statics.deleteAccount = async function (id) {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     throw new Error("Invalid user ID");
   }
+
+  const userData = await this.findOne({ _id: id });
+  if (!userData) {
+    throw new Error("User not found");
+  }
+  setImmediate(() => deletedAccountEmail(userData.email, userData.fullName));
+
   // delete associated workouts
   await Workout.deleteMany({ user_id: id });
 
@@ -330,34 +278,6 @@ userSchema.statics.deleteAccount = async function (id) {
   if (!user) {
     throw new Error("User not found");
   }
-
-  const transporter = nodemailer.createTransport({
-    service: "Gmail",
-    auth: {
-      user: process.env.EMAIL_USER,
-      pass: process.env.EMAIL_PASS,
-    },
-  });
-
-  const mailOptions = {
-    from: `"WorkoutX Support" <${process.env.EMAIL_USER}>`,
-    to: user.email,
-    replyTo: "harish.practicemail@gmail.com",
-    subject: "Your WorkoutX Account Has Been Deleted",
-    html: `
-      <div style="font-family: Arial, sans-serif; max-width: 500px; padding: 20px;">
-        <h2>Account Deletion Confirmation</h2>
-        <p>Hello ${user.fullName || "User"},</p>
-        <p>Your WorkoutX account has been successfully deleted as per your request.</p>
-        <p>We’re sorry to see you go! If this was a mistake or you change your mind, You are welcome to sign up again anytime.</p>
-        <p>If you didn’t request this, please contact our support team immediately.</p>
-        <p>— The WorkoutX Team</p>
-      </div>
-    `,
-  };
-
-  await transporter.sendMail(mailOptions);
-
   return { message: "Account deleted successfully.", user: user };
 };
